Always run `:verifyPlugin` as part of `check`
