+++ plugins/amazonq/build.gradle.kts
@@ -5,28 +5,15 @@ import software.aws.toolkits.gradle.intellij.IdeFlavor
 import software.aws.toolkits.gradle.intellij.IdeVersions
 
 plugins {
-    id("org.jetbrains.intellij")
+    id("toolkit-publishing-conventions")
 }
 
-val ideProfile = IdeVersions.ideProfile(project)
-
-val toolkitVersion: String by project
-val publishToken: String by project
-val publishChannel: String by project
-
-version = "$toolkitVersion-${ideProfile.shortName}"
-
 intellij {
-    version.set(ideProfile.community.version())
-    localPath.set(ideProfile.community.localPath())
     plugins.set(
         listOf(
             project(":plugin-core")
         )
     )
-
-    updateSinceUntilBuild.set(false)
-    instrumentCode.set(false)
 }
 
 dependencies {
@@ -35,20 +22,6 @@ dependencies {
     implementation(project(":plugin-amazonq:codewhisperer"))
     implementation(project(":plugin-amazonq:mynah-ui"))
     implementation(project(":plugin-amazonq:shared"))
-}
-
-configurations {
-    all {
-        // IDE provides netty
-        exclude("io.netty")
-    }
-
-    // Make sure we exclude stuff we either A) ships with IDE, B) we don't use to cut down on size
-    runtimeClasspath {
-        exclude(group = "org.slf4j")
-        exclude(group = "org.jetbrains.kotlin")
-        exclude(group = "org.jetbrains.kotlinx")
-    }
 }
 
 val moduleOnlyJar = tasks.create<Jar>("moduleOnlyJar") {
