+++ plugins/toolkit/intellij-standalone/src/main/resources/META-INF/plugin.xml
@@ -1,0 +1,5 @@
+<!-- Copyright 2024 Amazon.com, Inc. or its affiliates. All Rights Reserved. -->
+<!-- SPDX-License-Identifier: Apache-2.0 -->
+<idea-plugin xmlns:xi="http://www.w3.org/2001/XInclude">
+    <xi:include href="/META-INF/plugin-toolkit.xml" />
+</idea-plugin>
