+++ plugins/toolkit/jetbrains-core/resources/META-INF/plugin-toolkit.xml
