+++ plugins/toolkit/jetbrains-gateway/build.gradle.kts
@@ -155,3 +155,21 @@ tasks.integrationTest {
     val testToken = RandomString.make(32)
     environment("CWM_HOST_STATUS_OVER_HTTP_TOKEN", testToken)
 }
+
+configurations {
+    all {
+        // IDE provides netty
+        exclude("io.netty")
+    }
+
+    // Make sure we exclude stuff we either A) ships with IDE, B) we don't use to cut down on size
+    runtimeClasspath {
+        exclude(group = "org.slf4j")
+        exclude(group = "org.jetbrains.kotlin")
+        exclude(group = "org.jetbrains.kotlinx")
+    }
+}
+
+tasks.check {
+    dependsOn(tasks.verifyPlugin)
+}
